[gcc] Split reward classes into a separate module.
